# [flutter] No copy in ChangeNotifier.notifyListeners