Change Notifier has a fairly complicated contract that dictates how and when added/removed listeners are called. To accomplish this, it always copies a linked list of listeners on each notification cycle.

The efficiency of this method can be improved dramatically without adjusting the contract by using index based iteration (removing the potential concurrent modification exception). This is safe to do provided 1) listeners are always added to the end (true) and 2) removal of listeners does not modify the list.

Instead, notifyListeners checks if it encountered any "disposed" callbacks and removes them all at once. This means the callbacks may still exist in the listener set for N more frames, but are still eventually removed. This behavior too could be adjusted by making the callback field nullable.

All of the existing behavior tests pass, and the existing benchmarks should be improved.

also, never use linked lists.

Current:

```
I/flutter ( 9050): addListener (0 listeners): 276.0 ns per iteration
I/flutter ( 9050): removeListener (0 listeners): 437.0 ns per iteration
I/flutter ( 9050): notifyListener (0 listeners): 404.0 ns per iteration
I/flutter ( 9050): addListener (1 listeners): 193.0 ns per iteration
I/flutter ( 9050): removeListener (1 listeners): 667.0 ns per iteration
I/flutter ( 9050): notifyListener (1 listeners): 423.0 ns per iteration
I/flutter ( 9050): addListener (2 listeners): 174.0 ns per iteration
I/flutter ( 9050): removeListener (2 listeners): 1872.0 ns per iteration
I/flutter ( 9050): notifyListener (2 listeners): 441.0 ns per iteration
I/flutter ( 9050): addListener (3 listeners): 240.0 ns per iteration
I/flutter ( 9050): removeListener (3 listeners): 1160.0 ns per iteration
I/flutter ( 9050): notifyListener (3 listeners): 441.0 ns per iteration
I/flutter ( 9050): addListener (4 listeners): 274.0 ns per iteration
I/flutter ( 9050): removeListener (4 listeners): 4355.0 ns per iteration
I/flutter ( 9050): notifyListener (4 listeners): 555.0 ns per iteration
I/flutter ( 9050): addListener (5 listeners): 2206.0 ns per iteration
I/flutter ( 9050): removeListener (5 listeners): 3134.0 ns per iteration
I/flutter ( 9050): notifyListener (5 listeners): 623.0 ns per iteration
```


With this change:

```
I/flutter ( 8838): ================ FORMATTED ==============
I/flutter ( 8838): addListener (0 listeners): 168.0 ns per iteration
I/flutter ( 8838): removeListener (0 listeners): 264.0 ns per iteration
I/flutter ( 8838): notifyListener (0 listeners): 163.0 ns per iteration
I/flutter ( 8838): addListener (1 listeners): 170.0 ns per iteration
I/flutter ( 8838): removeListener (1 listeners): 264.0 ns per iteration
I/flutter ( 8838): notifyListener (1 listeners): 156.0 ns per iteration
I/flutter ( 8838): addListener (2 listeners): 265.0 ns per iteration
I/flutter ( 8838): removeListener (2 listeners): 1035.0 ns per iteration
I/flutter ( 8838): notifyListener (2 listeners): 174.0 ns per iteration
I/flutter ( 8838): addListener (3 listeners): 438.0 ns per iteration
I/flutter ( 8838): removeListener (3 listeners): 5436.0 ns per iteration
I/flutter ( 8838): notifyListener (3 listeners): 185.0 ns per iteration
I/flutter ( 8838): addListener (4 listeners): 232.0 ns per iteration
I/flutter ( 8838): removeListener (4 listeners): 1549.0 ns per iteration
I/flutter ( 8838): notifyListener (4 listeners): 186.0 ns per iteration
I/flutter ( 8838): addListener (5 listeners): 240.0 ns per iteration
I/flutter ( 8838): removeListener (5 listeners): 5121.0 ns per iteration
I/flutter ( 8838): notifyListener (5 listeners): 201.0 ns per iteration
```