@jonahwilliams 2021-02-02
I'm not sure how particular we want to be about the behavior of a disposed ChangeNotifier. I like making the listeners field final and non-nullable