@goderbauer 2021-02-02
How does the performance of this compare to the suggestion in https://github.com/flutter/flutter/pull/71947?