@jonahwilliams 2021-02-02
do we have a benchmark? mine is way simpler for starters