@jonahwilliams 2021-02-02
That one still allocates new lists fwiw