@goderbauer 2021-02-02
We have benchmarks, they were recently updated in https://github.com/flutter/flutter/pull/71986.