@jonahwilliams 2021-02-02
I don't really want to go down the road of figuring out a meaningful benchmark here. I'll just use my own CN implementation